Add files via upload
This Django-based project efficiently scrapes hyperlinks from web pages and stores them in a database. Utilizing requests and BeautifulSoup for scraping, it offers a seamless way to manage links via Django's ORM and a user-friendly interface. Perfect for those looking to learn web scraping and Django, the project is easy to extend and customize. Contributions are welcome!